Requested changes:
New method in Settings should be incubating.
Better location for new integration tests.
Don't provide 'provider(Callable)' helper method on Settings, as it
doesn't have much value.
