+++ subprojects/configuration-cache/src/integTest/groovy/org/gradle/configurationcache/ConfigurationCacheIntegrationTest.groovy
@@ -31,7 +31,6 @@ import org.gradle.process.ExecOperations
 import org.gradle.tooling.provider.model.ToolingModelBuilderRegistry
 import org.gradle.workers.WorkerExecutor
 import org.slf4j.Logger
-import spock.lang.Issue
 import spock.lang.Unroll
 
 import javax.inject.Inject
@@ -1114,46 +1113,5 @@ class ConfigurationCacheIntegrationTest extends AbstractConfigurationCacheIntegr
 
         then:
         outputContains("value = value")
-    }
-
-    @Issue("gradle/gradle#14465")
-    def "configuration is cacheable when providers are used in settings"() {
-
-        given:
-        settingsFile << """
-            providers.systemProperty("org.gradle.booleanProperty").forUseAtConfigurationTime().orElse(false).get()
-        """
-        configurationCacheRun "help", "-Dorg.gradle.booleanProperty=true"
-        def firstRunOutput = removeVfsLogOutput(result.normalizedOutput)
-            .replaceAll(/Calculating task graph as no configuration cache is available for tasks: help\n/, '')
-            .replaceAll(/Configuration cache entry stored.\n/, '')
-
-        when:
-        configurationCacheRun "help", "-Dorg.gradle.booleanProperty=true"
-        def secondRunOutput = removeVfsLogOutput(result.normalizedOutput)
-            .replaceAll(/Reusing configuration cache.\n/, '')
-            .replaceAll(/Configuration cache entry reused.\n/, '')
-
-        then:
-        firstRunOutput == secondRunOutput
-    }
-
-    @Issue("gradle/gradle#14465")
-    def "configuration cache is invalidated after property change when providers are used in settings"() {
-
-        given:
-        settingsFile << """
-            providers.systemProperty("org.gradle.booleanProperty").forUseAtConfigurationTime().orElse(false).get()
-        """
-        configurationCacheRun "help", "-Dorg.gradle.booleanProperty=true"
-
-        when:
-        configurationCacheRun "help", "-Dorg.gradle.booleanProperty=false"
-        def secondRunOutput = result.normalizedOutput
-        println(secondRunOutput)
-
-        then:
-        secondRunOutput.contains("Configuration cache entry stored.")
-        secondRunOutput.contains("Calculating task graph as configuration cache cannot be reused because system property 'org.gradle.booleanProperty' has changed.")
     }
 }
